Implemented ore rarity, separated the item and ore classes, fixed texture and item name errors, fixed platnium named as "platinum", ore spawn balanced to spawn much less ores than before and another minor fixes.

+++ integration-mc1-16/src/main/kotlin/org/eln2/utils/OreData.kt
@@ -8,6 +8,12 @@ import net.minecraft.item.ItemStack
  *
  * @param hardness How hard is the block (pickaxe type basically)
  * @param name Name of the ore (eg, "native_copper")
+ * @param rarity Rarity of the ore (eg, ->
+ * Very Rare 0 - 24
+ * Rare 25 - 49
+ * Uncommon 50 - 73
+ * Common 74 - 97
+ * Very Common 98 - 120)
  * @param result The item to drop
  */
-data class OreData(val hardness: Float, val name: String, val rarity: Float, val result: ItemStack)
+data class OreData(val hardness: Float, val name: String, val rarity: Int, val result: ItemStack)
